requirements.txt
basicsr>=1.4.2
boto3==1.26.69
# facexlib>=0.2.5
# gfpgan>=1.3.5
# numpy
# opencv-python
mysql-connector-python==8.0.32
numpy==1.21.6
opencv-python==4.7.0.68
pika==1.3.1
Pillow
python-dotenv==0.21.1
torch>=1.7
torchvision
tqdm